Renamed bob-server to bobserver
Because of reasons
